move testing summary
