+++ tests/integration/conftest.py
@@ -14,8 +14,8 @@ from starkware.starknet.public.abi import AbiType
 from typing_extensions import Protocol
 
 from protostar.commands.test.test_command import TestCommand
-from protostar.commands.test.testing_summary import TestingSummary
 from protostar.compiler.project_cairo_path_builder import ProjectCairoPathBuilder
+from protostar.test_runner import TestingSummary
 from protostar.utils.log_color_provider import LogColorProvider
 from tests.conftest import run_devnet
 from tests.integration.protostar_fixture import (
